Move shareable code out of runtime.
