+++ src/Shared/Hpack/EncoderHeaderEntry.cs
