+++ src/Shared/Hpack/README.md
@@ -1,0 +1,3 @@
+HPack dynamic compression. These files are kept separate to help avoid ASP.NET Core dependencies being added to them.
+
+Runtime currently doesn't implement HPack dynamic compression. These files will move into runtime shareable code in the future when support is added to runtime.
